+++ pytket/docs/conf.py
@@ -40,7 +40,7 @@ author = "Quantinuum"
 # The short X.Y version
 version = "1.31"
 # The full version, including alpha/beta/rc tags
-release = "1.31.0"
+release = "1.31.1"
 
 
 # -- General configuration ---------------------------------------------------