+++ pytket/docs/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+1.31.1 (August 2024)
+--------------------
+
+* Fix issue with css styles in pytket-circuit-renderer 0.9.
+
 1.31.0 (August 2024)
 --------------------